logic change to make mkv files when encoding 4k, or mp4 when encoding 1080p

+++ scripts/ripdisc.sh
@@ -214,14 +214,21 @@ fi
 # anything that is not 4k gets encoded qsv_264. qsv_265 used for 4k, simply because it saves disk space
 widthdigit=$(mediainfo "${newfile_name}" | grep ^Width | awk '{ print $3 }')
 case ${widthdigit} in
-    3) preset_import_file="$(dirname "$(readlink -f "$0")")/../presets/4k_qsv.json"; preset="4k_qsv" ;;
-    *) preset_import_file="$(dirname "$(readlink -f "$0")")/../presets/1080p_qsv.json"; preset="1080p_qsv" ;;
+    3) preset_import_file="$(dirname "$(readlink -f "$0")")/../presets/4k_qsv.json"; preset="4k_qsv"; extension="mkv" ;;
+    *) preset_import_file="$(dirname "$(readlink -f "$0")")/../presets/1080p_qsv.json"; preset="1080p_qsv"; extension="mp4" ;;
 esac
+
+output_basename=$(basename "$newfile" .mkv)
+output_file="${output_dir}/${output_basename}${testfilename}.${extension}"
+if [ -f "${output_file}" ]; then
+    echo "Target already exists: ${output_file} -- skipping."
+    continue
+fi
 
 # encode the file with HandBrakeCLI
 echo "Encoding with HandBrake (using ${encoder})..."
 log=$(mktemp -t handbrake.log.XXXX)
-flatpak run --command=HandBrakeCLI fr.handbrake.ghb --preset-import-file "${preset_import_file}" --preset "${preset}" -i "${newfile_name}" -o "${encode_dir}/${newfile}" 2> ${log}
+flatpak run --command=HandBrakeCLI fr.handbrake.ghb --preset-import-file "${preset_import_file}" --preset "${preset}" -i "${newfile_name}" -o "${output_file}" 2> ${log}
 if [ $? -eq 0 ]; then
     echo "HandBrake encode successful."
     rm -f ${log}
